+++ lib/Expr/ArrayExprOptimizer.cpp
@@ -7,16 +7,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "ArrayExprOptimizer.h"
-#include "ArrayExprRewriter.h"
-#include "ArrayExprVisitor.h"
-#include "AssignmentGenerator.h"
+#include "klee/Expr/ArrayExprOptimizer.h"
 
 #include "klee/Config/Version.h"
+#include "klee/Expr/ArrayExprRewriter.h"
+#include "klee/Expr/ArrayExprVisitor.h"
 #include "klee/Expr/Assignment.h"
+#include "klee/Expr/AssignmentGenerator.h"
 #include "klee/Expr/ExprBuilder.h"
-#include "klee/Support/ErrorHandling.h"
 #include "klee/OptionCategories.h"
+#include "klee/Support/ErrorHandling.h"
 #include "klee/util/BitArray.h"
 
 #include <llvm/ADT/APInt.h>
