+++ lib/Expr/ArrayExprVisitor.cpp
@@ -7,7 +7,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "ArrayExprVisitor.h"
+#include "klee/Expr/ArrayExprVisitor.h"
 
 #include "klee/Support/ErrorHandling.h"
