+++ lib/Expr/AssignmentGenerator.cpp
@@ -7,7 +7,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "AssignmentGenerator.h"
+#include "klee/Expr/AssignmentGenerator.h"
 
 #include "klee/Expr/Assignment.h"
 #include "klee/Support/ErrorHandling.h"
